move git fns to their own module
